Created a script and interface for damage.
I created the interface IDamageable that can be called when we want to
do damage to something. I also made a class that holds health but can be
modified for player stats in the future and have a seperate one for
enemies. As of right now if something with a damageSource touches
something with a health bar it subtracts health.